refactor: rename solver class
