+++ solver/solver.cpp
@@ -4,7 +4,7 @@
 #include <boost/numeric/ublas/matrix.hpp>
 #include <boost/numeric/ublas/vector.hpp>
 #include <boost/numeric/ublas/io.hpp>
-#include "Solver.h"
+#include "solver.h"
 #include "utils/tools.h"
