+++ solver/solver.h
